+++ migrations/capsules.sql
@@ -1,0 +1,39 @@
+INSERT INTO public.capsule
+(id, "content", created_at, sub_theme_capsule_id, title)
+VALUES(20, 'Si la version de toi à 8 ans te voyait aujourd''hui, qu''est-ce qu''elle trouverait cool dans ta vie ?', '2026-01-30 11:02:54.000', 1, 'Le regard des 8 ans');
+INSERT INTO public.capsule
+(id, "content", created_at, sub_theme_capsule_id, title)
+VALUES(21, 'Qu''as-tu arrêté de faire “parce que ce n''était pas sérieux”, alors que ça te faisait vibrer ?', '2026-01-30 11:02:54.000', 1, 'Le sérieux toxique');
+INSERT INTO public.capsule
+(id, "content", created_at, sub_theme_capsule_id, title)
+VALUES(22, 'À quel moment t''es-tu éloignée de ce qui te rendait légère ?', '2026-01-30 11:02:54.000', 1, 'La légèreté');
+INSERT INTO public.capsule
+(id, "content", created_at, sub_theme_capsule_id, title)
+VALUES(23, 'Qu''est-ce qui te faisait te sentir vivante avant que tout devienne une to-do list ?', '2026-01-30 11:02:54.000', 1, 'La To-do list vs La Vie');
+INSERT INTO public.capsule
+(id, "content", created_at, sub_theme_capsule_id, title)
+VALUES(24, 'Quelle part de toi essaies-tu de cacher pour rester “acceptable” ?', '2026-01-30 11:02:54.000', 2, 'L''image acceptable');
+INSERT INTO public.capsule
+(id, "content", created_at, sub_theme_capsule_id, title)
+VALUES(25, 'Dans quel domaine joues-tu un rôle qui te coûte plus qu''il ne t''apporte ?', '2026-01-30 11:02:54.000', 2, 'Le rôle épuisant');
+INSERT INTO public.capsule
+(id, "content", created_at, sub_theme_capsule_id, title)
+VALUES(26, 'Si tu étais honnête à 100 %, qu''est-ce que tu dirais différemment ?', '2026-01-30 11:02:54.000', 2, 'L''honnêteté 100%');
+INSERT INTO public.capsule
+(id, "content", created_at, sub_theme_capsule_id, title)
+VALUES(27, 'Quelle vérité intérieure attends-tu encore d''avoir le droit d''écouter ?', '2026-01-30 11:02:54.000', 2, 'La permission');
+INSERT INTO public.capsule
+(id, "content", created_at, sub_theme_capsule_id, title)
+VALUES(28, 'Qu''est-ce que tu essaies de sauver dans un chapitre qui est déjà terminé ?', '2026-01-30 11:02:54.000', 3, 'Le chapitre terminé');
+INSERT INTO public.capsule
+(id, "content", created_at, sub_theme_capsule_id, title)
+VALUES(29, 'Si tu devais laisser une seule peur derrière la porte de ton ancienne vie, laquelle serait-ce ?', '2026-01-30 11:02:54.000', 3, 'Le bagage');
+INSERT INTO public.capsule
+(id, "content", created_at, sub_theme_capsule_id, title)
+VALUES(30, 'Pourquoi as-tu peur du vide entre deux chapitres, alors que c''est là que tout s''invente ?', '2026-01-30 11:02:54.000', 3, 'La peur du vide');
+INSERT INTO public.capsule
+(id, "content", created_at, sub_theme_capsule_id, title)
+VALUES(31, 'Si tu repartais de zéro demain, sans aucune obligation, que ferais-tu différemment ?', '2026-01-30 11:02:54.000', 3, 'La page blanche');
+INSERT INTO public.capsule
+(id, "content", created_at, sub_theme_capsule_id, title)
+VALUES(32, 'Qu''est-ce que la "nouvelle toi" sait déjà, mais que "l''ancienne toi" refuse d''entendre ?', '2026-01-30 11:02:54.000', 3, 'L''ancienne vs la nouvelle');