+++ migrations/sub_theme_capsule.sql
@@ -1,0 +1,9 @@
+INSERT INTO public.sub_theme_capsule
+(id, code, "name", theme_capsule_id)
+VALUES(1, 'INNER_CHILD', 'Enfant intérieur & joie', 1);
+INSERT INTO public.sub_theme_capsule
+(id, code, "name", theme_capsule_id)
+VALUES(2, 'SHADOW_SELF', 'L''ombre & l''authenticité', 1);
+INSERT INTO public.sub_theme_capsule
+(id, code, "name", theme_capsule_id)
+VALUES(3, 'LIFE_TRANSITIONS', 'Transitions de vie', 1);